fix: update validation
